[NVMESH-3537] Fix Clang 14 build errors on Ubuntu 22.04
The numeric_downcast macro's statement expression was not
preserving the narrowed type through to the caller, leading to
format string mismatches when used with printf-style functions.
Wrapping the statement expression result in an explicit cast to
the target type ensures the macro reliably returns the expected
type, which satisfies Clang's format string checking.

Several format specifiers were also corrected to match their
actual argument types: the status field uses %d instead of %hhu
since the underlying type is a signed int, and pass2toma is
explicitly cast to u8 to match its %hhu specifier. The op field
is similarly cast to u8 to align with the format string.

Finally, the jdr_object_scope and jdr_array_scope macros declare
RAII-style cleanup variables that are never referenced by name.
Clang 14 warns about these unused variables, so the unused
attribute is added alongside the existing cleanup attribute to
suppress the false positive.

+++ clnt/block/datapath_ec/recov/nvmeibc_block_dp_ec_recov_hot.c
@@ -1093,7 +1093,7 @@ static void __nvmeibc_free_jrnl_ents_request_pet_describe(struct htr_ctx *h, int
 		&h->so->o->journal,
 		"free_jrnl_ents.request(sgmnt_idx=%hhu, blkset_slba=%llu, blkset_num=%llu, pass2toma=%hhu, lock_id=0x%llx<union nvmeib_lock_id>)",
 		numeric_downcast(u8, si), free_ents_comp->blkset_slba, free_ents_comp->blkset_num,
-		free_ents_comp->pass2toma, free_ents_comp->lock_ent);
+		(u8)free_ents_comp->pass2toma, free_ents_comp->lock_ent);
 }
 
 static void __nvmeibc_send_recovered_response_pet_describe(struct htr_ctx *h, int si, int rv)
@@ -1103,12 +1103,12 @@ static void __nvmeibc_send_recovered_response_pet_describe(struct htr_ctx *h, in
 	if (h->tx_jentries[si].is_valid) {
 		NVMEIBC_IO_PET_MSG_NORM(
 			&h->so->o->journal,
-			"free_jrnl_ents.response(sgmnt_idx=%hhu, status=%hhu) = %d",
+			"free_jrnl_ents.response(sgmnt_idx=%hhu, status=%d) = %d",
 			numeric_downcast(u8, si), gen_cmd->rsp.br.status, rv);
 	} else {
 		NVMEIBC_IO_PET_MSG_NORM(
 			&h->so->o->journal,
-			"send_recovered_blkset.response(sgmnt_idx=%hhu, status=%hhu) = %d",
+			"send_recovered_blkset.response(sgmnt_idx=%hhu, status=%d) = %d",
 			numeric_downcast(u8, si), gen_cmd->rsp.br.status, rv);
 	}
 }
